refactored scoring system from GridVM --> small VMs
