--- JasmineTests/Common/Support/GridViewModelMocks.swift
@@ -1,44 +1,0 @@
-import Foundation
-@testable import Jasmine
-
-class GridViewModelSameRowMock: GridViewModel {
-    override var scoreOnCorrectLine: Int {
-        return 10
-    }
-
-    override var scoreOnWon: Int {
-        return 1_000
-    }
-
-    override func lineIsCorrect(_ line: [Coordinate]) -> Bool {
-        return line.map { $0.row }.isAllSame
-    }
-}
-
-class GridViewModelSameColumnMock: GridViewModel {
-    override var scoreOnCorrectLine: Int {
-        return 10
-    }
-
-    override var scoreOnWon: Int {
-        return 1_000
-    }
-
-    override func lineIsCorrect(_ line: [Coordinate]) -> Bool {
-        return line.map { $0.col }.isAllSame
-    }
-}
-
-class GridViewModelFirstRowMock: GridViewModel {
-    override var scoreOnCorrectLine: Int {
-        return 10
-    }
-
-    override var scoreOnWon: Int {
-        return 1_000
-    }
-
-    override func lineIsCorrect(_ line: [Coordinate]) -> Bool {
-        return line.isAllTrue { $0.row == 0 }
-    }
-}
